requirements.txt
beautifulsoup4==4.10.0
pandas==1.3.5
ndjson==0.3.1
apache-airflow==2.2.4
apache-airflow-providers-amazon==3.0.0
openpyxl==3.0.9
